Added define so we can easily switch to loading address from EEPROM, also defined team directions

+++ atmega/inc/diff_drive.h
@@ -34,7 +34,7 @@ void dd_disable(dd *rob);
 void dd_comm_test(dd *rob);
 
 bool dd_is_loc(dd*rob , float posThresh);
-void dd_goto_spiral(dd *rob, double veloDes);
+void dd_goto_spiral(dd *rob, float veloDes);
 bool dd_goto_rot_trans(dd *rob, double veloDes);
 void dd_drive(dd *rob);
 void get_fault_status(dd *rob);
